refactor env generators
