+++ packages/core/src/context/provider.ts
@@ -1,5 +1,5 @@
 import type { RnvContext } from './types';
 
-export const getContext = (): RnvContext => {
+export const getContext = <C = any>(): RnvContext<C> => {
     return global.RNV_CONTEXT;
 };
