+++ packages/core/src/plugins/index.ts
@@ -15,11 +15,11 @@ import {
     sanitizeDynamicProps,
 } from '../system/fs';
 import { chalk, logDebug, logError, logInfo, logSuccess, logTask, logWarning } from '../logger';
-import { doResolve, doResolvePath } from '../system/resolve';
+import { doResolve } from '../system/resolve';
 import { RnvContext } from '../context/types';
 import { PluginCallback, RnvPlugin, RnvPluginScope } from './types';
 import { RenativeConfigPaths, RenativeConfigPlugin } from '../schema/types';
-import { RnvModuleConfig, RnvPlatform } from '../types';
+import { RnvPlatform } from '../types';
 import { inquirerPrompt } from '../api';
 import { writeRenativeConfigFile } from '../configs/utils';
 import { installPackageDependencies } from '../projects/npm';
@@ -934,70 +934,6 @@ export const getLocalRenativePlugin = () => ({
         },
     },
 });
-
-export const getModuleConfigs = (c: RnvContext): RnvModuleConfig => {
-    let modulePaths: Array<string> = [];
-    const moduleAliases: Record<string, string | undefined> = {};
-
-    const doNotResolveModulePaths: Array<string> = [];
-
-    // PLUGINS
-    parsePlugins(
-        c,
-        c.platform,
-        (plugin, pluginPlat, key) => {
-            const { webpackConfig } = plugin;
-
-            if (webpackConfig) {
-                if (webpackConfig.modulePaths) {
-                    if (typeof webpackConfig.modulePaths === 'boolean') {
-                        if (webpackConfig.modulePaths) {
-                            modulePaths.push(`node_modules/${key}`);
-                        }
-                    } else {
-                        webpackConfig.modulePaths.forEach((v) => {
-                            modulePaths.push(v);
-                        });
-                    }
-                }
-                const wpMa = webpackConfig.moduleAliases;
-                if (wpMa) {
-                    if (typeof wpMa === 'boolean') {
-                        moduleAliases[key] = doResolvePath(key, true, {}, c.paths.project.nodeModulesDir);
-                    } else {
-                        Object.keys(wpMa).forEach((aKey) => {
-                            const mAlias = wpMa[aKey];
-                            if (typeof mAlias === 'string') {
-                                moduleAliases[key] = doResolvePath(mAlias, true, {}, c.paths.project.nodeModulesDir);
-                                // DEPRECATED use => projectPath
-                                // } else if (mAlias.path) {
-                                //     moduleAliases[key] = path.join(c.paths.project.dir, mAlias.path);
-                            } else if (includesPluginPath(mAlias.projectPath)) {
-                                moduleAliases[key] = sanitizePluginPath(mAlias.projectPath, key);
-                            } else if (mAlias.projectPath) {
-                                moduleAliases[key] = path.join(c.paths.project.dir, mAlias.projectPath);
-                            }
-                        });
-                    }
-                }
-            }
-        },
-        true
-    );
-
-    const moduleAliasesArray: Array<string> = [];
-    Object.keys(moduleAliases).forEach((key) => {
-        moduleAliasesArray.push(`${key}:${moduleAliases[key]}`);
-    });
-
-    modulePaths = modulePaths
-        .map((v) => v && doResolvePath(v, true, {}, c.paths.project.dir)!)
-        .concat(doNotResolveModulePaths)
-        .concat([c.paths.project.assets.dir])
-        .filter(Boolean);
-
-    return { modulePaths, moduleAliases, moduleAliasesArray };
-};
 
 export const updateRenativeConfigs = async (c: RnvContext) => {
     await loadPluginTemplates(c);
